apply migration in serve_helper
this is not ideal but the easiest thing to do rn
