+++ fastn/Cargo.toml
@@ -10,8 +10,6 @@ homepage.workspace = true
 [dependencies]
 clap.workspace = true
 colored.workspace = true
-diesel = { workspace = true, optional = true }
-diesel_migrations = { workspace = true, optional = true }
 fastn-observer.workspace = true
 fastn-update.workspace = true
 fastn-core.workspace = true
@@ -27,4 +25,4 @@ fastn-ds.workspace = true
 
 
 [features]
-auth = ["fastn-core/auth", "diesel", "diesel_migrations"]
+auth = ["fastn-core/auth"]
